refactor(money): move Quick Buy A/B config into QuickBuy folder (TSA-612)
Relocate the socialAiTSA612AbtestQuickBuy A/B test config out of
TokenDetails/components/abTestConfig.ts into a new abTestConfig.ts under
the existing QuickBuy folder, since it's a Quick Buy concern rather than
TokenDetails-specific. The unrelated Ambient Price Color test stays in
TokenDetails. Importers (TokenDetails view/test, MarketInsightsView,
abTestAnalyticsRegistry) repointed accordingly.
