+++ app/util/analytics/abTestAnalyticsRegistry.ts
@@ -9,10 +9,8 @@ import {
   HUB_PAGE_DISCOVERY_TABS_AB_TEST_ANALYTICS_MAPPING,
   WALLET_HOME_POST_ONBOARDING_AB_TEST_ANALYTICS_MAPPING,
 } from '../../components/Views/Homepage/abTestConfig';
-import {
-  AMBIENT_PRICE_COLOR_AB_TEST_ANALYTICS_MAPPING,
-  SOCIAL_AI_QUICK_BUY_AB_TEST_ANALYTICS_MAPPING,
-} from '../../components/UI/TokenDetails/components/abTestConfig';
+import { AMBIENT_PRICE_COLOR_AB_TEST_ANALYTICS_MAPPING } from '../../components/UI/TokenDetails/components/abTestConfig';
+import { SOCIAL_AI_QUICK_BUY_AB_TEST_ANALYTICS_MAPPING } from '../../components/Views/SocialLeaderboard/TraderPositionView/components/QuickBuy/abTestConfig';
 import { WHATS_HAPPENING_EXPLORE_AB_TEST_ANALYTICS_MAPPING } from '../../components/Views/TrendingView/abTestConfig';
 import { ONBOARDING_CHECKLIST_STEPPER_AB_TEST_ANALYTICS_MAPPING } from '../../components/UI/WalletHomeOnboardingSteps/abTestConfig';
 
@@ -39,5 +37,7 @@ export const AB_TEST_ANALYTICS_MAPPINGS: readonly ABTestAnalyticsMapping[] = [
 
   // Token Details
   AMBIENT_PRICE_COLOR_AB_TEST_ANALYTICS_MAPPING,
+
+  // Quick Buy
   SOCIAL_AI_QUICK_BUY_AB_TEST_ANALYTICS_MAPPING,
 ];
